refactor: move fonts and springs to constants
