+++ src/client/components/alerts/alert.tsx
@@ -9,11 +9,11 @@ import { Outline } from "client/components/ui/outline";
 import { ReactiveButton } from "client/components/ui/reactive-button";
 import { Shadow } from "client/components/ui/shadow";
 import { Text } from "client/components/ui/text";
+import { fonts } from "client/constants/fonts";
+import { springs } from "client/constants/springs";
 import { useMotion, useRem } from "client/hooks";
 import { Alert, selectAlertIndex } from "client/store/alert";
 import { selectIsMenuOpen } from "client/store/menu";
-import { fonts } from "client/utils/fonts";
-import { springs } from "client/utils/springs";
 import { images, playSound, sounds } from "shared/assets";
 import { palette } from "shared/constants/palette";
 import { brightenIfDark, darken } from "shared/utils/color-utils";
