+++ src/client/components/menu/home/home-footer.tsx
@@ -1,7 +1,7 @@
 import Roact from "@rbxts/roact";
 import { Text } from "client/components/ui/text";
+import { fonts } from "client/constants/fonts";
 import { useRem } from "client/hooks";
-import { fonts } from "client/utils/fonts";
 import { palette } from "shared/constants/palette";
 
 export function HomeFooter() {
