+++ src/client/components/menu/home/home-version.tsx
@@ -1,7 +1,7 @@
 import Roact from "@rbxts/roact";
 import { Text } from "client/components/ui/text";
+import { fonts } from "client/constants/fonts";
 import { useRem } from "client/hooks";
-import { fonts } from "client/utils/fonts";
 import { $git, $package } from "rbxts-transform-debug";
 import { palette } from "shared/constants/palette";
