+++ src/client/components/menu/skins/act-button.tsx
@@ -6,11 +6,11 @@ import { sendAlert } from "client/alerts";
 import { PrimaryButton } from "client/components/ui/primary-button";
 import { Shadow } from "client/components/ui/shadow";
 import { Text } from "client/components/ui/text";
+import { fonts } from "client/constants/fonts";
+import { springs } from "client/constants/springs";
 import { useMotion, useRem } from "client/hooks";
 import { selectMenuCurrentSkin } from "client/store/menu";
-import { fonts } from "client/utils/fonts";
 import { formatInteger } from "client/utils/format-integer";
-import { springs } from "client/utils/springs";
 import { sounds } from "shared/assets";
 import { USER_NAME } from "shared/constants/core";
 import { palette } from "shared/constants/palette";
