+++ src/client/components/menu/support/support-heart.tsx
@@ -2,8 +2,8 @@ import { lerpBinding, useInterval } from "@rbxts/pretty-react-hooks";
 import Roact, { useEffect } from "@rbxts/roact";
 import { setTimeout } from "@rbxts/set-timeout";
 import { Image } from "client/components/ui/image";
+import { springs } from "client/constants/springs";
 import { useMotion, useRem } from "client/hooks";
-import { springs } from "client/utils/springs";
 import { images } from "shared/assets";
 
 export function SupportHeart() {
