+++ src/client/components/stats/stats-card.tsx
@@ -5,9 +5,9 @@ import { Group } from "client/components/ui/group";
 import { ReactiveButton } from "client/components/ui/reactive-button";
 import { Shadow } from "client/components/ui/shadow";
 import { Text } from "client/components/ui/text";
+import { fonts } from "client/constants/fonts";
+import { springs } from "client/constants/springs";
 import { useMotion, useRem } from "client/hooks";
-import { fonts } from "client/utils/fonts";
-import { springs } from "client/utils/springs";
 import { palette } from "shared/constants/palette";
 
 interface StatsCardProps {
