+++ src/client/components/ui/text-field.tsx
@@ -1,6 +1,6 @@
 import Roact, { useEffect, useState } from "@rbxts/roact";
 
-import { fonts } from "../../utils/fonts";
+import { fonts } from "../../constants/fonts";
 import { Group } from "./group";
 import { TextProps } from "./text";
