+++ src/client/components/world/candy/candy-item.tsx
@@ -3,8 +3,8 @@ import { composeBindings } from "@rbxts/pretty-react-hooks";
 import Roact, { memo, useEffect, useMemo } from "@rbxts/roact";
 import { Image } from "client/components/ui/image";
 import { Shadow } from "client/components/ui/shadow";
+import { springs } from "client/constants/springs";
 import { useMotion, useRem, useSeed } from "client/hooks";
-import { springs } from "client/utils/springs";
 import { images } from "shared/assets";
 import { CandyType } from "shared/store/candy";
 import { brighten } from "shared/utils/color-utils";
