+++ src/client/components/world/candy/candy.tsx
@@ -1,9 +1,9 @@
 import { useSelector } from "@rbxts/react-reflex";
 import Roact, { useBinding, useEffect, useMemo } from "@rbxts/roact";
 import { Group } from "client/components/ui/group";
+import { springs } from "client/constants/springs";
 import { useMotion, useRem } from "client/hooks";
 import { selectWorldCamera } from "client/store/world";
-import { springs } from "client/utils/springs";
 
 import { CandyItem } from "./candy-item";
 import { useCandyOnScreen } from "./use-candy-on-screen";
