+++ src/client/components/world/snakes/snake-head.tsx
@@ -3,9 +3,9 @@ import Roact, { useEffect, useMemo, useState } from "@rbxts/roact";
 import { setTimeout } from "@rbxts/set-timeout";
 import { Group } from "client/components/ui/group";
 import { Image } from "client/components/ui/image";
+import { springs } from "client/constants/springs";
 import { useContinuousAngle, useMotion, useRem } from "client/hooks";
 import { selectWorldInputAngle } from "client/store/world";
-import { springs } from "client/utils/springs";
 import { getSnakeSkin, getSnakeSkinForTracer } from "shared/constants/skins";
 import { subtractRadians } from "shared/utils/math-utils";
