+++ src/client/components/world/snakes/snakes.tsx
@@ -1,9 +1,9 @@
 import { useSelector } from "@rbxts/react-reflex";
 import Roact, { useEffect, useMemo } from "@rbxts/roact";
 import { Group } from "client/components/ui/group";
+import { springs } from "client/constants/springs";
 import { useMotion, useRem } from "client/hooks";
 import { selectWorldCamera } from "client/store/world";
-import { springs } from "client/utils/springs";
 
 import { Snake } from "./snake";
 import { useSnakesOnScreen } from "./use-snakes-on-screen";
