+++ src/client/components/world/snakes/use-snake-bindings.ts
@@ -2,7 +2,7 @@ import { useSelectorCreator } from "@rbxts/react-reflex";
 import { createMotion, immediate, spring } from "@rbxts/ripple";
 import Roact, { createBinding, useEffect, useMemo, useRef } from "@rbxts/roact";
 import { RunService } from "@rbxts/services";
-import { springs } from "client/utils/springs";
+import { springs } from "client/constants/springs";
 import { describeSnakeFromScore, selectSnakeIsBoosting, SnakeEntity } from "shared/store/snakes";
 
 import { SnakeOnScreen } from "./use-snakes-on-screen";
