+++ src/client/components/world/world-border.tsx
@@ -4,9 +4,9 @@ import { CanvasGroup } from "client/components/ui/canvas-group";
 import { Frame } from "client/components/ui/frame";
 import { Group } from "client/components/ui/group";
 import { Image } from "client/components/ui/image";
+import { springs } from "client/constants/springs";
 import { useMotion, useRem } from "client/hooks";
 import { selectWorldCamera } from "client/store/world";
-import { springs } from "client/utils/springs";
 import { images } from "shared/assets";
 import { WORLD_BOUNDS } from "shared/constants/core";
 import { palette } from "shared/constants/palette";
